Remove auto-update from example service files
In the current solution the Jenkins and docker-registry services
automatically updates when restarted. This is because the Jenkins and
docker-registry example service files track a moving targets
in their `ExecStart` definition and there are `ExecStartPre` steps
removing and pulling the target image before running.

This commit removes this behavior and forces all updates to be done
explicitly by performing the steps eariler done in `ExecStartPre`
manually. The rational is that there are cases where services need to be
restarted but it is not suitable to at the same time update the version
of them, since that might cause breakages.

Signed-off-by: Viktor Sjölind <vsjolind@luxoft.com>